fix: make #dropdown react to new  content
